Refactor many commands into consistency
