+++ src/commands/audit-log/get-audit-log.ts
@@ -1,0 +1,86 @@
+import { Separator, select } from '@socketsecurity/registry/lib/prompts'
+import { Spinner } from '@socketsecurity/registry/lib/spinner'
+
+import {
+  handleApiCall,
+  handleUnsuccessfulApiResponse
+} from '../../utils/api.ts'
+import { setupSdk } from '../../utils/sdk.ts'
+
+type Choice<Value> = {
+  description?: string
+  disabled?: boolean | string
+  name?: string
+  type?: never
+  value: Value
+}
+
+type AuditChoice = Choice<string>
+
+type AuditChoices = (Separator | AuditChoice)[]
+
+export async function getAuditLog({
+  apiToken,
+  orgSlug,
+  outputJson,
+  outputMarkdown,
+  page,
+  perPage,
+  type
+}: {
+  apiToken: string
+  outputJson: boolean
+  outputMarkdown: boolean
+  orgSlug: string
+  page: number
+  perPage: number
+  type: string
+}): Promise<void> {
+  const spinner = new Spinner({
+    text: `Looking up audit log for ${orgSlug}\n`
+  }).start()
+
+  const socketSdk = await setupSdk(apiToken)
+  const result = await handleApiCall(
+    socketSdk.getAuditLogEvents(orgSlug, {
+      outputJson,
+      outputMarkdown,
+      orgSlug,
+      type,
+      page,
+      per_page: perPage
+    }),
+    `Looking up audit log for ${orgSlug}\n`
+  )
+
+  if (!result.success) {
+    handleUnsuccessfulApiResponse('getAuditLogEvents', result, spinner)
+    return
+  }
+
+  spinner.stop()
+
+  const data: AuditChoices = []
+  const logDetails: { [key: string]: string } = {}
+
+  for (const d of result.data.results) {
+    const { created_at } = d
+    if (created_at) {
+      const name = `${new Date(created_at).toLocaleDateString('en-us', { year: 'numeric', month: 'numeric', day: 'numeric' })} - ${d.user_email} - ${d.type} - ${d.ip_address} - ${d.user_agent}`
+      data.push(<AuditChoice>{ name }, new Separator())
+      logDetails[name] = JSON.stringify(d.payload)
+    }
+  }
+
+  console.log(
+    logDetails[
+      (await select({
+        message: type
+          ? `\n Audit log for: ${orgSlug} with type: ${type}\n`
+          : `\n Audit log for: ${orgSlug}\n`,
+        choices: data,
+        pageSize: 30
+      })) as any
+    ]
+  )
+}
