+++ src/commands/repos/cmd-repos-delete.ts
@@ -1,0 +1,63 @@
+import meowOrDie from 'meow'
+import colors from 'yoctocolors-cjs'
+
+import { deleteRepo } from './delete-repo.ts'
+import { AuthError } from '../../utils/errors'
+import { getDefaultToken } from '../../utils/sdk'
+
+import type { CliCommandConfig } from '../../utils/meow-with-subcommands'
+
+const config: CliCommandConfig = {
+  commandName: 'delete',
+  description: 'Delete a repository in an organization',
+  hidden: false,
+  flags: {},
+  help: (parentName, config) => `
+    Usage
+      $ ${parentName} ${config.commandName} <org slug> <repo slug>
+
+    Examples
+      $ ${parentName} ${config.commandName} FakeOrg test-repo
+  `
+}
+
+export const cmdReposDelete = {
+  description: config.description,
+  hidden: config.hidden,
+  run
+}
+
+async function run(
+  argv: readonly string[],
+  importMeta: ImportMeta,
+  { parentName }: { parentName: string }
+): Promise<void> {
+  const cli = meowOrDie(config.help(parentName, config), {
+    argv,
+    description: config.description,
+    importMeta,
+    flags: config.flags,
+    allowUnknownFlags: false
+  })
+
+  const [orgSlug = '', repoName = ''] = cli.input
+
+  if (!orgSlug || !repoName) {
+    console.error(`${colors.bgRed(colors.white('Input error'))}: Please provide the required fields:\n
+      - Org name as the first argument ${!orgSlug ? colors.red('(missing!)') : colors.green('(ok)')}\n
+      - Repository name as the second argument ${!repoName ? colors.red('(missing!)') : typeof repoName !== 'string' ? colors.red('(invalid!)') : colors.green('(ok)')}\n
+      - At least one TARGET (e.g. \`.\` or \`./package.json\`
+    `)
+    cli.showHelp()
+    return
+  }
+
+  const apiToken = getDefaultToken()
+  if (!apiToken) {
+    throw new AuthError(
+      'User must be authenticated to run this command. To log in, run the command `socket login` and enter your API key.'
+    )
+  }
+
+  await deleteRepo(orgSlug, repoName, apiToken)
+}
