+++ src/commands/repos/cmd-repos-update.ts
@@ -1,0 +1,112 @@
+import meowOrDie from 'meow'
+import colors from 'yoctocolors-cjs'
+
+import { updateRepo } from './update-repo.ts'
+import { commonFlags, outputFlags } from '../../flags'
+import { AuthError } from '../../utils/errors'
+import { getFlagListOutput } from '../../utils/output-formatting'
+import { getDefaultToken } from '../../utils/sdk'
+
+import type { CliCommandConfig } from '../../utils/meow-with-subcommands'
+
+const config: CliCommandConfig = {
+  commandName: 'create',
+  description: 'Update a repository in an organization',
+  hidden: false,
+  flags: {
+    ...commonFlags,
+    ...outputFlags,
+    repoName: {
+      type: 'string',
+      shortFlag: 'n',
+      default: '',
+      description: 'Repository name'
+    },
+    repoDescription: {
+      type: 'string',
+      shortFlag: 'd',
+      default: '',
+      description: 'Repository description'
+    },
+    homepage: {
+      type: 'string',
+      shortFlag: 'h',
+      default: '',
+      description: 'Repository url'
+    },
+    defaultBranch: {
+      type: 'string',
+      shortFlag: 'b',
+      default: 'main',
+      description: 'Repository default branch'
+    },
+    visibility: {
+      type: 'string',
+      shortFlag: 'v',
+      default: 'private',
+      description: 'Repository visibility (Default Private)'
+    }
+  },
+  help: (parentName, config) => `
+    Usage
+      $ ${parentName} ${config.commandName} <org slug>
+
+    Options
+      ${getFlagListOutput(config.flags, 6)}
+
+    Examples
+      $ ${parentName} ${config.commandName} FakeOrg
+  `
+}
+
+export const cmdReposUpdate = {
+  description: config.description,
+  hidden: config.hidden,
+  run
+}
+
+async function run(
+  argv: readonly string[],
+  importMeta: ImportMeta,
+  { parentName }: { parentName: string }
+): Promise<void> {
+  const cli = meowOrDie(config.help(parentName, config), {
+    argv,
+    description: config.description,
+    importMeta,
+    flags: config.flags,
+    allowUnknownFlags: false
+  })
+
+  const repoName = cli.flags['repoName']
+  const [orgSlug = ''] = cli.input
+
+  if (!repoName || typeof repoName !== 'string' || !orgSlug) {
+    console.error(`${colors.bgRed(colors.white('Input error'))}: Please provide the required fields:\n
+      - Org name as the first argument ${!orgSlug ? colors.red('(missing!)') : colors.green('(ok)')}\n
+      - Repository name using --repoName ${!repoName ? colors.red('(missing!)') : typeof repoName !== 'string' ? colors.red('(invalid!)') : colors.green('(ok)')}\n
+      - At least one TARGET (e.g. \`.\` or \`./package.json\`
+    `)
+    cli.showHelp()
+    return
+  }
+
+  const apiToken = getDefaultToken()
+  if (!apiToken) {
+    throw new AuthError(
+      'User must be authenticated to run this command. To log in, run the command `socket login` and enter your API key.'
+    )
+  }
+
+  await updateRepo({
+    apiToken,
+    outputJson: Boolean(cli.flags['json']),
+    outputMarkdown: Boolean(cli.flags['markdown']),
+    orgSlug,
+    repoName,
+    description: String(cli.flags['repoDescription'] || ''),
+    homepage: String(cli.flags['homepage'] || ''),
+    default_branch: String(cli.flags['defaultBranch'] || ''),
+    visibility: String(cli.flags['visibility'] || 'private')
+  })
+}
