+++ src/commands/repos/cmd-repos.ts
@@ -1,24 +1,24 @@
-import { create } from './create'
-import { del } from './delete'
-import { list } from './list'
-import { update } from './update'
-import { view } from './view'
+import { cmdReposCreate } from './cmd-repos-create.ts'
+import { cmdReposDelete } from './cmd-repos-delete.ts'
+import { cmdReposList } from './cmd-repos-list.ts'
+import { cmdReposUpdate } from './cmd-repos-update.ts'
+import { cmdReposView } from './cmd-repos-view.ts'
 import { meowWithSubcommands } from '../../utils/meow-with-subcommands'
 
 import type { CliSubcommand } from '../../utils/meow-with-subcommands'
 
 const description = 'Repositories related commands'
 
-export const reposCommand: CliSubcommand = {
+export const cmdRepos: CliSubcommand = {
   description,
   async run(argv, importMeta, { parentName }) {
     await meowWithSubcommands(
       {
-        create,
-        view,
-        list,
-        del,
-        update
+        cmdReposCreate,
+        cmdReposView,
+        cmdReposList,
+        cmdReposDelete,
+        cmdReposUpdate
       },
       {
         argv,
