+++ src/commands/scan/cmd-scan-create.ts
