+++ src/commands/scan/cmd-scan-delete.ts
