+++ src/commands/scan/cmd-scan-list.ts
