+++ src/commands/scan/cmd-scan-metadata.ts
