+++ src/commands/scan/cmd-scan-stream.ts
