--- src/commands/types.ts
@@ -1,7 +1,0 @@
-export interface CommandContext {
-  scope: string
-  time: number
-  repo: string
-  outputJson: boolean
-  file: string
-}
